tests/test-requirements.txt
avro-python3==1.7.7
nose==1.3.7
PyHamcrest==1.8.5
